xhrSetup should extract withCredentials too, be called with url argument and thus from within loadInternal + we should move helper func to utils
