+++ lib/integration/p2p-loader-generator.js
@@ -1,31 +1,6 @@
 import SegmentView from './mapping/segment-view';
 import TrackView from './mapping/track-view';
-
-import {BaseXHR} from 'xhr-shaper';
-
-function extendHeadersFromXhrSetup(xhrSetup, headersBase) {
-    var headers = headersBase || {};
-
-    // this will get called when the `on...` event handler properties
-    // are set on BaseXHR
-    var noEventHandlers = () => {
-        throw new Error('Can\'t use event handler property in this xhrSetup function. Contact Streamroot support.');
-    };
-
-    const xhrMock = new BaseXHR({
-        setRequestHeader: (header, value) => {
-            headers[header] = value;
-        }
-    }, noEventHandlers, noEventHandlers, noEventHandlers);
-
-    // In case xhrSetup calls functions others than the one implemented
-    // above on xhrMock (like `setRequestHeader`) then BaseXHR will throw
-    // We will also throw explicitly at an attempt to set event handlers (on...)
-    // A call to an undefined function will throw anyway
-    xhrSetup && xhrSetup(xhrMock);
-
-    return headers;
-}
+import {extractInfoFromXhrSetup} from '../utils';
 
 /**
  * Generates a P2PLoader class definition that will use the Streamroot Hls.js wrapper instance passed
@@ -44,9 +19,9 @@ const P2PLoaderGenerator = function (hlsjsWrapper) {
 
         constructor(config) {
             if (config) {
-                // IMPORTANT: these are custom config parameters that are not present in the original hls.js configuration
-                this.withCredentials = !!(config.request && config.request.withCredentials);
-                this.headers = extendHeadersFromXhrSetup(config.xhrSetup, config.request && config.request.headers ? config.request.headers : {});
+                this.xhrSetup = config.xhrSetup;
+                // initialize with basic header config. we might extend this as we evaluate xhrSetup
+                this.headers = config.request && config.request.headers ? config.request.headers : {};
             }
             this.reset();
         }
@@ -157,6 +132,11 @@ const P2PLoaderGenerator = function (hlsjsWrapper) {
             if (this.peerAgentLoader) {
                 throw new Error('P2P loader was not reset correctly, internal state indicates unfinalized request');
             }
+
+            let {headers, withCredentials} = extractInfoFromXhrSetup(this.xhrSetup, this.url, this.headers);
+
+            this.headers = headers;
+            this.withCredentials = withCredentials;
 
             if (this.byteRange) {
                 this.headers.Range = `bytes=${this.frag.byteRangeStartOffset}-${this.frag.byteRangeEndOffset - 1}`;
