+++ lib/utils.js
@@ -1,3 +1,5 @@
+import {BaseXHR} from 'xhr-shaper';
+
 function inheritStaticPropertiesReadOnly(target, source) {
 
     function defineReadOnlyProperty(staticProperty) {
@@ -16,6 +18,32 @@ function inheritStaticPropertiesReadOnly(target, source) {
     }
 }
 
+function extractInfoFromXhrSetup(xhrSetup, url, headersBase) {
+    var headers = headersBase || {};
+
+    // this will get called when the `on...` event handler properties
+    // are set on BaseXHR
+    var noEventHandlers = () => {
+        throw new Error('Can\'t use event handler property in this xhrSetup function. Contact Streamroot support.');
+    };
+
+    const xhrMock = new BaseXHR({
+        setRequestHeader: (header, value) => {
+            headers[header] = value;
+        },
+        withCredentials: false
+    }, noEventHandlers, noEventHandlers, noEventHandlers);
+
+    // In case xhrSetup calls functions others than the one implemented
+    // above on xhrMock (like `setRequestHeader`) then BaseXHR will throw
+    // We will also throw explicitly at an attempt to set event handlers (on...)
+    // A call to an undefined function will throw anyway
+    xhrSetup && xhrSetup(xhrMock, url);
+
+    return {headers, withCredentials: xhrMock.withCredentials};
+}
+
 export {
-    inheritStaticPropertiesReadOnly
+    inheritStaticPropertiesReadOnly,
+    extractInfoFromXhrSetup
 };
